Move robot event subscribers
